+++ tests/testthat/test-clean.R
@@ -1,5 +1,71 @@
+test_that("Count", {
+  num <- sample(1:10, 15, TRUE) # Create matrix
+  mtx <- matrix(data = num, nrow = 5, ncol = 3)
+
+  mtx[c(2, 4, 13)] <- NA # Add missing values
+
+  ## Count missing values in rows
+  expect_equal(count(mtx, f = is.na, margin = 1), c(0, 1, 1, 1, 0))
+  ## Count non-missing values in columns
+  no_na <- count(mtx, f = is.na, margin = 2, negate = TRUE)
+  expect_equal(no_na, c(3, 5, 4))
+
+  df <- as.data.frame(mtx)
+
+  ## Count missing values in rows
+  expect_equal(count(df, f = is.na, margin = 1), c(0, 1, 1, 1, 0))
+  ## Count non-missing values in columns
+  no_na <- count(df, f = is.na, margin = 2, negate = TRUE)
+  expect_equal(no_na, c(V1 = 3, V2 = 5, V3 = 4))
+})
+test_that("Detect", {
+  num <- sample(1:10, 15, TRUE) # Create matrix
+  mtx <- matrix(data = num, nrow = 5, ncol = 3)
+
+  mtx[c(2, 4, 13)] <- NA # Add missing values
+
+  ## Find row with NA
+  expect_equal(detect(mtx, f = is.na, margin = 1), c(FALSE, TRUE, TRUE, TRUE, FALSE))
+  ## Find column without any NA
+  no_na <- detect(mtx, f = is.na, margin = 2, negate = TRUE, all = TRUE)
+  expect_equal(no_na, c(FALSE, TRUE, FALSE))
+
+  df <- as.data.frame(mtx)
+
+  ## Find row with NA
+  expect_equal(detect(df, f = is.na, margin = 1), c(FALSE, TRUE, TRUE, TRUE, FALSE))
+  ## Find column without any NA
+  no_na <- detect(df, f = is.na, margin = 2, negate = TRUE, all = TRUE)
+  expect_equal(no_na, c(V1 = FALSE, V2 = TRUE, V3 = FALSE))
+})
+test_that("Compact", {
+  num <- sample(1:10, 25, TRUE) # Create matrix
+  mtx <- matrix(data = num, nrow = 5, ncol = 5)
+
+  # Nothing to remove
+  clean <- compact(mtx, f = is_zero, margin = 1, all = FALSE)
+  expect_equal(dim(clean), dim(mtx))
+
+  mtx[1, ] <- 0 # Add zeros
+
+  # Nothing to remove
+  clean <- compact(mtx, f = is_zero, margin = 2, all = TRUE)
+  expect_equal(dim(clean), dim(mtx))
+
+  mtx[, 1] <- 0 # Add zeros
+
+  # Remove rows
+  clean_row <- compact(mtx, f = is_zero, margin = 1, all = TRUE)
+  expect_lt(nrow(clean_row), nrow(mtx))
+  expect_equal(ncol(clean_row), ncol(mtx))
+
+  # Remove columns
+  clean_col <- compact(mtx, f = is_zero, margin = 2, all = TRUE)
+  expect_lt(ncol(clean_col), ncol(mtx))
+  expect_equal(nrow(clean_col), nrow(mtx))
+})
+# Missing values ===============================================================
 test_that("Replace missing values", {
-  options(arkhe.verbose = FALSE)
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -8,15 +74,8 @@ test_that("Replace missing values", {
   # Replace NAs
   clean <- replace_NA(mtx, value = 999)
   expect_equal(sum(clean == 999), 3)
-
-  ct <- as_count(mtx) # Create CountMatrix
-
-  # Replace NAs
-  ct_clean <- replace_NA(ct, value = 999L)
-  expect_equal(sum(ct_clean == 999), 3)
 })
 test_that("Remove missing values", {
-  options(arkhe.verbose = FALSE)
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -29,31 +88,17 @@ test_that("Remove missing values", {
   # Remove rows
   clean_row <- remove_NA(mtx, margin = 1)
   expect_lt(nrow(clean_row), nrow(mtx))
+  expect_gt(nrow(clean_row), 0)
   expect_equal(ncol(clean_row), ncol(mtx))
 
   # Remove columns
-  clean_row <- remove_NA(mtx, margin = 2)
-  expect_lt(ncol(clean_row), ncol(mtx))
-  expect_equal(nrow(clean_row), nrow(mtx))
-
-  ct <- as_count(mtx) # Create CountMatrix
-
-  # Remove rows
-  ct_clean_row <- remove_NA(ct, margin = 1)
-  expect_lt(nrow(ct_clean_row), nrow(ct))
-  expect_equal(ncol(ct_clean_row), ncol(ct))
-
-  # Remove columns
-  ct_clean_col <- remove_NA(ct, margin = 2)
-  expect_lt(ncol(ct_clean_col), ncol(ct))
-  expect_equal(nrow(ct_clean_col), nrow(ct))
-
-  # Check message
-  options(arkhe.verbose = TRUE)
-  expect_message(remove_NA(mtx, margin = 1))
+  clean_col <- remove_NA(mtx, margin = 2)
+  expect_lt(ncol(clean_col), ncol(mtx))
+  expect_gt(ncol(clean_col), 0)
+  expect_equal(nrow(clean_col), nrow(mtx))
 })
+# Infinite values ==============================================================
 test_that("Replace infinite values", {
-  options(arkhe.verbose = FALSE)
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -64,7 +109,6 @@ test_that("Replace infinite values", {
   expect_equal(sum(clean == 999), 3)
 })
 test_that("Remove infinite values", {
-  options(arkhe.verbose = FALSE)
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -77,19 +121,17 @@ test_that("Remove infinite values", {
   # Remove rows
   clean_row <- remove_Inf(mtx, margin = 1)
   expect_lt(nrow(clean_row), nrow(mtx))
+  expect_gt(nrow(clean_row), 0)
   expect_equal(ncol(clean_row), ncol(mtx))
 
   # Remove columns
-  clean_row <- remove_Inf(mtx, margin = 2)
-  expect_lt(ncol(clean_row), ncol(mtx))
-  expect_equal(nrow(clean_row), nrow(mtx))
-
-  # Check message
-  options(arkhe.verbose = TRUE)
-  expect_message(remove_Inf(mtx, margin = 1))
+  clean_col <- remove_Inf(mtx, margin = 2)
+  expect_lt(ncol(clean_col), ncol(mtx))
+  expect_gt(ncol(clean_col), 0)
+  expect_equal(nrow(clean_col), nrow(mtx))
 })
+# Zeros ========================================================================
 test_that("Replace zeros", {
-  options(arkhe.verbose = FALSE)
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -98,15 +140,8 @@ test_that("Replace zeros", {
   # Replace 0s
   clean <- replace_zero(mtx, value = 999)
   expect_equal(sum(clean == 999), 3)
-
-  ct <- as_count(mtx) # Create CountMatrix
-
-  # Replace 0s
-  ct_clean <- replace_zero(ct, value = 999L)
-  expect_equal(sum(ct_clean == 999), 3)
 })
 test_that("Remove zeros", {
-  options(arkhe.verbose = FALSE)
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -119,31 +154,17 @@ test_that("Remove zeros", {
   # Remove rows
   clean_row <- remove_zero(mtx, margin = 1)
   expect_lt(nrow(clean_row), nrow(mtx))
+  expect_gt(nrow(clean_row), 0)
   expect_equal(ncol(clean_row), ncol(mtx))
 
   # Remove columns
-  clean_row <- remove_zero(mtx, margin = 2)
-  expect_lt(ncol(clean_row), ncol(mtx))
-  expect_equal(nrow(clean_row), nrow(mtx))
-
-  ct <- as_count(mtx) # Create CountMatrix
-
-  # Remove rows
-  ct_clean_row <- remove_zero(ct, margin = 1)
-  expect_lt(nrow(ct_clean_row), nrow(ct))
-  expect_equal(ncol(ct_clean_row), ncol(ct))
-
-  # Remove columns
-  ct_clean_col <- remove_zero(ct, margin = 2)
-  expect_lt(ncol(ct_clean_col), ncol(ct))
-  expect_equal(nrow(ct_clean_col), nrow(ct))
-
-  # Check message
-  options(arkhe.verbose = TRUE)
-  expect_message(remove_zero(mtx, margin = 1))
+  clean_col <- remove_zero(mtx, margin = 2)
+  expect_lt(ncol(clean_col), ncol(mtx))
+  expect_gt(ncol(clean_col), 0)
+  expect_equal(nrow(clean_col), nrow(mtx))
 })
-test_that("Remove empty row/column (numeric)", {
-  options(arkhe.verbose = FALSE)
+# Empty rows/columns ===========================================================
+test_that("Remove empty rows/columns (numeric)", {
   num <- sample(1:10, 25, TRUE) # Create matrix
   mtx <- matrix(data = num, nrow = 5, ncol = 5)
 
@@ -156,32 +177,15 @@ test_that("Remove empty row/column (numeric)", {
 
   # Remove rows
   clean_row <- remove_empty(mtx, margin = 1)
-  expect_lt(nrow(clean_row), nrow(mtx))
+  expect_equal(nrow(clean_row), 4)
   expect_equal(ncol(clean_row), ncol(mtx))
 
   # Remove columns
   clean_col <- remove_empty(mtx, margin = 2)
-  expect_lt(ncol(clean_col), ncol(mtx))
+  expect_equal(ncol(clean_col), 4)
   expect_equal(nrow(clean_col), nrow(mtx))
-
-  ct <- as_count(mtx) # Create CountMatrix
-
-  # Remove rows
-  ct_clean_row <- remove_empty(ct, margin = 1)
-  expect_lt(nrow(ct_clean_row), nrow(ct))
-  expect_equal(ncol(ct_clean_row), ncol(ct))
-
-  # Remove columns
-  ct_clean_col <- remove_empty(ct, margin = 2)
-  expect_lt(ncol(ct_clean_col), ncol(ct))
-  expect_equal(nrow(ct_clean_col), nrow(ct))
-
-  # Check message
-  options(arkhe.verbose = TRUE)
-  expect_message(remove_empty(mtx, margin = 1))
 })
-test_that("Remove empty row/column (character)", {
-  options(arkhe.verbose = FALSE)
+test_that("Remove empty rows/columns (character)", {
   char <- sample(LETTERS, 25, TRUE) # Create matrix
   mtx <- matrix(data = char, nrow = 5, ncol = 5)
 
@@ -194,19 +198,32 @@ test_that("Remove empty row/column (character)", {
 
   # Remove rows
   clean_row <- remove_empty(mtx, margin = 1)
-  expect_lt(nrow(clean_row), nrow(mtx))
+  expect_equal(nrow(clean_row), 4)
   expect_equal(ncol(clean_row), ncol(mtx))
 
   # Remove columns
   clean_col <- remove_empty(mtx, margin = 2)
-  expect_lt(ncol(clean_col), ncol(mtx))
+  expect_equal(ncol(clean_col), 4)
   expect_equal(nrow(clean_col), nrow(mtx))
 })
-test_that("Remove empty row/column (logical)", {
-  options(arkhe.verbose = FALSE)
+test_that("Remove empty rows/columns (logical)", {
   char <- sample(c(TRUE, FALSE), 25, TRUE) # Create matrix
   mtx <- matrix(data = char, nrow = 5, ncol = 5)
 
   # Nothing to remove
-  expect_error(remove_empty(mtx, margin = 1))
+  clean <- remove_empty(mtx, margin = 1)
+  expect_equal(dim(clean), dim(mtx))
+
+  mtx[1, ] <- NA # Add NA
+  mtx[, 1] <- NA # Add NA
+
+  # Remove rows
+  clean_row <- remove_empty(mtx, margin = 1)
+  expect_equal(nrow(clean_row), 4)
+  expect_equal(ncol(clean_row), ncol(mtx))
+
+  # Remove columns
+  clean_col <- remove_empty(mtx, margin = 2)
+  expect_equal(ncol(clean_col), 4)
+  expect_equal(nrow(clean_col), nrow(mtx))
 })