+++ tests/testthat/test-utilities.R
@@ -1,16 +1,3 @@
-test_that("Compact", {
-  expect_length(compact(is.null, list("A", NULL, "B")), 2)
-})
-test_that("Detect", {
-  k <- detect(is.na, c(1, 2, NA, 4, 5, NA))
-  expect_type(k, "logical")
-  expect_equal(sum(k), 2)
-})
-test_that("Count", {
-  k <- count(is.na, c(1, 2, NA, 4, 5, NA))
-  expect_type(k, "integer")
-  expect_equal(k, 2)
-})
 test_that("Extract", {
   k <- extract(c("abc123", "def456", "ghi"), "[1-9]{3}")
   expect_type(k, "character")